Updated example for a Vec of Rule
